+++ br/pkg/lightning/backend/external/BUILD.bazel
@@ -1,0 +1,24 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "external",
+    srcs = [
+        "file.go",
+        "sharedisk.go",
+    ],
+    importpath = "github.com/pingcap/tidb/br/pkg/lightning/backend/external",
+    visibility = ["//visibility:public"],
+    deps = ["//br/pkg/storage"],
+)
+
+go_test(
+    name = "external_test",
+    timeout = "short",
+    srcs = ["file_test.go"],
+    embed = [":external"],
+    flaky = True,
+    deps = [
+        "//br/pkg/storage",
+        "@com_github_stretchr_testify//require",
+    ],
+)
