+++ br/pkg/lightning/backend/external/file.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package sharedisk
+package external
 
 import (
 	"context"
